refactor(runtime): move runtime specific types to their own file
